vehicle_search_api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
mysql-connector-python
python-dotenv
pydantic
pytest
requests
pytest-mock
